Move some common code out of TypeInference.kt
